+++ backend/api/views/exception_handler.py
@@ -22,4 +22,24 @@ def api_exception_handler(exc, context):
         logger.warning("exception", class_name=exc.__class__.__name__, exception=exc)
         return exc.response()
 
-    return views.exception_handler(exc, context)
+    return rename_detail_to_message(views.exception_handler(exc, context))
+
+
+def rename_detail_to_message(response):
+    """
+    rename DRF's "detail" field to "message", to match our own style
+    """
+
+    if not hasattr(response, "data") or "detail" not in response.data:
+        return response
+
+    if "message" not in response.data:
+        response.data["message"] = response.data["detail"]
+        del response.data["detail"]
+    else:
+        # not sure what to do in this case, let's just exchange them
+        detail = response.data["message"]
+        response.data["message"] = response.data["detail"]
+        response.data["detail"] = detail
+
+    return response