+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Increase the number of tasks displayable in frontend workflow [#697](https://github.com/Substra/substra-backend/pull/697)
+- Increase the number of tasks displayable in frontend workflow ([#697](https://github.com/Substra/substra-backend/pull/697))
+- BREAKING: Change the format of some error API responses from `{"detail": ...}` to `{"message": ...}`, to match the rest of the API ([#699](https://github.com/Substra/substra-backend/pull/699))
 
 ## [0.40.0](https://github.com/Substra/substra-backend/releases/tag/0.40.0) 2023-07-25